# NDEV-2059: Improve performance of get_account_rooted_slot, get_branch_slots, get_sol_sig_rooted_slot