Since most queries for get_account_rooted_slot result in either the last rooted_slot or adjacent slots, the first query should execute on a single day (input slot - 216000 slots)

This removes the fullscan on the update_slot table for a lot of requests.

Original with full scan:
```
SELECT DISTINCT slot
FROM events.update_account_distributed
WHERE (pubkey = [238, 205, 99, 85, 74, 252, 237, 43, 151, 11, 212, 155, 237, 78, 226, 30, 144, 127, 0, 158, 237, 154, 182, 209, 81, 245, 77, 5, 91, 181, 49, 210]) AND (slot <= 240920250) AND (slot IN (
    SELECT slot
    FROM events.update_slot
    WHERE status = 'Rooted'
    ORDER BY slot DESC
))
ORDER BY slot DESC
LIMIT 1

Query id: 90185cca-e051-4af4-9971-ed495332bcfc

┌──────slot─┐
│ 240920250 │
└───────────┘

1 row in set. Elapsed: 4.245 sec. Processed 45.31 million rows, 397.17 MB (10.67 million rows/s., 93.57 MB/s.)
```


New one:
```
SELECT DISTINCT slot
FROM events.update_account_distributed
WHERE (pubkey = [238, 205, 99, 85, 74, 252, 237, 43, 151, 11, 212, 155, 237, 78, 226, 30, 144, 127, 0, 158, 237, 154, 182, 209, 81, 245, 77, 5, 91, 181, 49, 210]) AND (slot <= 240920250) AND (slot IN (
    SELECT slot
    FROM events.update_slot
    WHERE (status = 'Rooted') AND (slot >= GREATEST(0, 240920250 - 216000))
))
ORDER BY slot DESC
LIMIT 1

Query id: ef49ce36-0739-4a6c-ba72-38367fdaf63d

┌──────slot─┐
│ 240920250 │
└───────────┘

1 row in set. Elapsed: 0.092 sec. Processed 933.37 thousand rows, 8.64 MB (10.18 million rows/s., 94.23 MB/s.)
```