@Deniskore 2023-09-02
I've introduced a new table called rooted_slots in pull request https://github.com/neonlabsorg/geyser-neon-filter/pull/45, which focuses exclusively on rooted slots. Leveraging this table has led to significant performance improvements in queries.